requirements.txt
pandas==1.3.4
streamlit==1.2.0
rdkit-pypi==2021.9.5.1